Add BVG API client
Background
==========
Add BVG API client and a rudimentary chat bot that interacts with the
client. The chat bot can print 10 stations that match the search
criteria or print the next two departures for a given station. It can
print a help message that shows all the commands it understands.

The application should handle unknown commands and client errors.

There is room for improvement about the command parsing and probably we
should add a debounce on the send button. But I left these for future
iterations.